requirements_pg.txt
pandas==2.2.2
openpyxl==3.1.5
pydantic==2.8.2
tqdm==4.66.5
pdfplumber==0.11.4
python-docx==1.1.2
langchain-text-splitters==0.3.0
python-dotenv==1.0.1
numpy==1.26.4
openai==1.40.2
psycopg[binary]==3.2.1
pgvector==0.2.5
tiktoken==0.7.0